complete refactor add landing page vs app
