+++ docs/tree.md
@@ -3,43 +3,20 @@
 │   ├── localhost-key.pem
 │   └── localhost.pem
 ├── components.json
-├── docs
-│   ├── about.md
-│   ├── mission.md
-│   ├── PortfolioPage.md
-│   ├── TradePage.md
-│   ├── tree.md
-│   ├── UPDATES.md
-│   └── ZeroDevMigration.md
 ├── eslint.config.mjs
 ├── next.config.ts
 ├── next-env.d.ts
 ├── package.json
 ├── package-lock.json
 ├── postcss.config.mjs
-├── public
-│   ├── file.svg
-│   ├── globe.svg
-│   ├── next.svg
-│   ├── SmartPortfolioMobile.png
-│   ├── SP_LOGO.png
-│   ├── SP_ONLY_ICON.png
-│   ├── SP_ONLY_LETTERS.png
-│   ├── sw.js
-│   ├── vercel.svg
-│   └── window.svg
 ├── README.md
 ├── src
 │   ├── abi
 │   │   └── erc20Faucet.ts
-│   ├── app
-│   │   ├── actions.ts
+│   ├── (app)
 │   │   ├── activity
 │   │   │   └── page.tsx
-│   │   ├── favicon.ico
-│   │   ├── globals.css
 │   │   ├── layout.tsx
-│   │   ├── manifest.ts
 │   │   ├── page.tsx
 │   │   ├── rewards
 │   │   │   └── page.tsx
@@ -49,16 +26,26 @@
 │   │       ├── components
 │   │       │   └── CreatePortfolio.tsx
 │   │       └── page.tsx
+│   ├── app
+│   │   ├── actions.ts
+│   │   ├── favicon.ico
+│   │   ├── globals.css
+│   │   ├── layout.tsx
+│   │   ├── manifest.ts
+│   │   └── page.tsx
 │   ├── components
 │   │   ├── BottomNav.tsx
 │   │   ├── Complex_deprecated_PortfolioChart.tsx
+│   │   ├── DebugInfo.tsx
 │   │   ├── Faucet.tsx
 │   │   ├── Header.tsx
+│   │   ├── InstallPrompt.tsx
 │   │   ├── LoadingAnimation.tsx
 │   │   ├── MotionWrapper.tsx
 │   │   ├── PageWrapper.tsx
 │   │   ├── PortfolioChart.tsx
 │   │   ├── providers.tsx
+│   │   ├── PushNotificationManager.tsx
 │   │   ├── TokenBalanceDisplay.tsx
 │   │   └── ui
 │   │       ├── alert.tsx
@@ -72,8 +59,11 @@
 │   │       ├── select.tsx
 │   │       ├── skeleton.tsx
 │   │       ├── slider.tsx
+│   │       ├── sonner.tsx
 │   │       ├── table.tsx
 │   │       └── tabs.tsx
+│   ├── config
+│   │   └── zerodev.ts
 │   ├── contexts
 │   │   ├── AccountContext.tsx
 │   │   ├── PortfolioContext.tsx
@@ -82,8 +72,11 @@
 │   ├── contracts
 │   │   └── addresses.json
 │   └── lib
-│       └── utils.ts
+│       ├── passkey.ts
+│       ├── utils.ts
+│       ├── web3.ts
+│       └── zerodev.ts
 ├── tailwind.config.ts
 └── tsconfig.json
 
-16 directories, 70 files
+16 directories, 63 files
